ZIP bugs behoben Options hinzugefügt (werden als String mit dazugehörigem Textfeld angezeigt)
testklassen gelöscht
abstrakte klasse und interface für die widgets erweitert

widgets an die erweiterungen angepasst

+++ src/de/szut/SmartGadgetBar/Widgets/ZIP/Decomprimator.java
@@ -2,6 +2,7 @@ package de.szut.SmartGadgetBar.Widgets.ZIP;
 
 import java.io.BufferedInputStream;
 import java.io.BufferedOutputStream;
+import java.io.File;
 import java.io.FileInputStream;
 import java.io.FileOutputStream;
 import java.util.zip.ZipEntry;
@@ -9,7 +10,8 @@ import java.util.zip.ZipInputStream;
 
 /**
  * 
- * @author Fabian Klasse zum deomprimieren von *.zip Dateien
+ * Klasse zum dekomprimieren von *.zip Dateien
+ * Sollte ueber die Klasse ZIP benutzt werden
  */
 public class Decomprimator {
 	final static int BUFFER = 2048;
@@ -22,16 +24,26 @@ public class Decomprimator {
 					new BufferedInputStream(fis));
 			ZipEntry entry;
 			while ((entry = zis.getNextEntry()) != null) {
+				File f = new File(entry.getName());
+				System.out.println(f.getAbsolutePath());
+				if(entry.isDirectory()){
+					f.mkdirs();
+				}else{
+					f.createNewFile();
+				}
+				
 				int count;
 				byte data[] = new byte[BUFFER];
-				// write the files to the disk
-				FileOutputStream fos = new FileOutputStream(entry.getName());
-				dest = new BufferedOutputStream(fos, BUFFER);
-				while ((count = zis.read(data, 0, BUFFER)) != -1) {
-					dest.write(data, 0, count);
+				if(f.isFile()){
+					FileOutputStream fos = new FileOutputStream(entry.getName());
+					dest = new BufferedOutputStream(fos, BUFFER);
+					while ((count = zis.read(data, 0, BUFFER)) != -1) {
+						dest.write(data, 0, count);
+					}
+					dest.flush();
+					dest.close();
+					
 				}
-				dest.flush();
-				dest.close();
 			}
 			zis.close();
 		} catch (Exception e) {
